.github/scripts/requirements.txt
lxml==4.8.0
translate-toolkit==3.5.2
